requirements.txt
pandas>=2.0,<3.0
numpy>=1.24,<3.0
matplotlib>=3.7,<4.0
plotly>=5.15,<6.0
requests>=2.28,<3.0
python-dotenv>=1.0,<2.0
statsmodels>=0.14,<1.0
pyarrow>=14.0,<20.0
fredapi>=0.5,<1.0
wrds>=3.2,<4.0
